Check is first login after sign in
